Note Commit
Three mode of operation should be available, all activated based on #defines or arguments passed to start_tsi() and are mutually exclusive:

 1. Hardware polling: The TSI module continually scans the indicated pin registers, and the TSI0_GENCS[ESOR] pin is set to 1 to allow for End-of-Scan interrupts that, when triggered, check if the values from the enable TSI pin registers are sane, and copy them to a buffer array, which is made available for use.
2. Software polling: The TSI module can be started, and once started, scans, and then triggers an End-of-Scan interrupt that copies the values from the result registers to a buffer (but only if they're not out of range), and then triggers another scan. I'm not sure if there's any advantage that this method has over the hardware polling mode.
3. Software triggered: A function is made available that, when called, returns a pointer to a buffer array of pin results, and either doesn't return until all pins selected have been measured, or has an End-of-Scan interrupt that copies to the buffer and returns a pointer to the buffer immediately, as opposed to waiting for all the pins to be scanned before copying the results.

All modes will have a start_tsi() function, and it's function signature will probably look something like:

    enum MODES {HPOLL, SPOLL, STRIG};

    void start_tsi(something or_other, ..., MODES(uint8_t mode) );
    void start_tsi(something or_other, ..., ((MODES)(uint8_t mode)));

A variable has to be passed to the function to indicate which pins are going to be used, and therefore which pins the TSI module needs to scan. I think some macros should be defined, as well, for helping set up a value to be passed to the start_tsi() function, that is written directly into the TSI0_PEN register (after some casting to the right size).

For mode 2 and 3 (SPOLL and STRIG), a start_scan() function will be defined, which can optionally take a list of pins to be scanned, or will simply trigger a scan of the pins indicated in TSI0_PEN[0-15]. It'll probably be defined something like:

    #define start_scan {TSI0_GENCS |= TSI_GENCS_SWTS; wait(79);}

Lastly, for debugging purposes, the print_tsi_register_values() function should be included, and expanded to allow compile-time choosing of what it prints out.